--- src/main/java/com/abhinav/covid19/vaccinetracker/bot/BotService.java
@@ -1,7 +1,0 @@
-package com.abhinav.covid19.vaccinetracker.bot;
-
-import com.abhinav.covid19.vaccinetracker.model.VaccineCenters;
-
-public interface BotService {
-    void notify(String userId, VaccineCenters vaccineCenters);
-}
