--- src/main/java/com/abhinav/covid19/vaccinetracker/cowin/CowinConfig.java
@@ -1,13 +1,0 @@
-package com.abhinav.covid19.vaccinetracker.cowin;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-import lombok.Data;
-
-@Configuration
-@ConfigurationProperties("cowin")
-@Data
-public class CowinConfig {
-    private String apiUrl;
-}
