--- src/main/java/com/abhinav/covid19/vaccinetracker/model/UserRequest.java
@@ -1,9 +1,0 @@
-package com.abhinav.covid19.vaccinetracker.model;
-
-import lombok.Value;
-
-@Value
-public class UserRequest {
-    private String chatId;
-    private String pincode;
-}
