--- src/main/java/com/abhinav/covid19/vaccinetracker/persistence/VaccinePersistence.java
@@ -1,9 +1,0 @@
-package com.abhinav.covid19.vaccinetracker.persistence;
-
-import com.abhinav.covid19.vaccinetracker.model.VaccineCenters;
-
-public interface VaccinePersistence {
-    void persistVaccineCenters(String pincode, VaccineCenters vaccineCenters);
-
-    VaccineCenters fetchVaccineCentersByPincode(String pincode);
-}
