+++ src/main/java/org/covid19/vaccinetracker/bot/BotService.java
@@ -1,0 +1,7 @@
+package org.covid19.vaccinetracker.bot;
+
+import org.covid19.vaccinetracker.model.VaccineCenters;
+
+public interface BotService {
+    void notify(String userId, VaccineCenters vaccineCenters);
+}
