+++ src/main/java/org/covid19/vaccinetracker/model/UserRequest.java
@@ -1,0 +1,9 @@
+package org.covid19.vaccinetracker.model;
+
+import lombok.Value;
+
+@Value
+public class UserRequest {
+    String chatId;
+    String pincode;
+}
