+++ src/main/java/org/covid19/vaccinetracker/model/VaccineCenters.java
@@ -1,0 +1,17 @@
+package org.covid19.vaccinetracker.model;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
+
+import lombok.Data;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Data
+public class VaccineCenters {
+
+    @JsonProperty("centers")
+    public List<Center> centers = null;
+
+}
