+++ src/main/java/org/covid19/vaccinetracker/tracker/VaccineTracker.java
@@ -1,0 +1,33 @@
+package org.covid19.vaccinetracker.tracker;
+
+import org.covid19.vaccinetracker.cowin.CowinApiClient;
+import org.covid19.vaccinetracker.model.VaccineCenters;
+import org.covid19.vaccinetracker.persistence.VaccinePersistence;
+
+import org.springframework.stereotype.Service;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Service
+public class VaccineTracker {
+    private final CowinApiClient cowinApiClient;
+    private final VaccinePersistence vaccinePersistence;
+
+    public VaccineTracker(CowinApiClient cowinApiClient, VaccinePersistence vaccinePersistence) {
+        this.cowinApiClient = cowinApiClient;
+        this.vaccinePersistence = vaccinePersistence;
+    }
+
+    public void refreshVaccineAvailability(String pincode) {
+        log.info("Fetching vaccine availability for pincode {}", pincode);
+        final VaccineCenters vaccineCenters = cowinApiClient.fetchCentersByPincode(pincode);
+        log.info("Persisting vaccine availability for pincode {}", pincode);
+        vaccinePersistence.persistVaccineCenters(pincode, vaccineCenters);
+    }
+
+    public VaccineCenters fetchVaccineAvailability(String pincode) {
+        log.info("Fetching Vaccine availability from local store");
+        return vaccinePersistence.fetchVaccineCentersByPincode(pincode);
+    }
+}
