--- src/test/java/com/abhinav/covid19/vaccinetracker/Covid19VaccineTrackerApplicationTests.java
@@ -1,14 +1,0 @@
-package com.abhinav.covid19.vaccinetracker;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Covid19VaccineTrackerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-
-}
