+++ src/test/java/org/covid19/vaccinetracker/Covid19VaccineTrackerApplicationTests.java
@@ -1,0 +1,14 @@
+package org.covid19.vaccinetracker;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Covid19VaccineTrackerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+
+}
