+++ src/test/java/org/covid19/vaccinetracker/utils/UtilsTest.java
@@ -1,0 +1,20 @@
+package org.covid19.vaccinetracker.utils;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class UtilsTest {
+
+    @Test
+    public void pincodeValidationTest() {
+        assertTrue(Utils.isValidPincode("440022"), "Should be valid!");
+        assertTrue(Utils.isValidPincode("400008"), "Should be valid!");
+        assertTrue(Utils.isValidPincode("410038"), "Should be valid!");
+        assertFalse(Utils.isValidPincode("4561091"), "Should not be valid!");
+        assertFalse(Utils.isValidPincode("40010"), "Should not be valid!");
+        assertFalse(Utils.isValidPincode("abcde"), "Should not be valid!");
+        assertFalse(Utils.isValidPincode("395def"), "Should not be valid!");
+    }
+}
